+++ packages/nodes/table/src/transforms/index.ts
@@ -3,12 +3,12 @@
  */
 
 export * from './addColumn';
-export * from './insertTableRow';
-export * from './insertTableColumn';
 export * from './addRow';
 export * from './deleteColumn';
 export * from './deleteRow';
 export * from './deleteTable';
 export * from './insertTable';
+export * from './insertTableColumn';
+export * from './insertTableRow';
 export * from './moveSelectionFromCell';
 export * from './setTableColSize';
