+++ packages/ui/dnd/src/hooks/useDndBlock.ts
@@ -1,79 +1,12 @@
-import { useState } from 'react';
-import { getEmptyImage } from 'react-dnd-html5-backend';
-import { useEditorRef } from '@udecode/plate-core';
-import { DropLineDirection } from '../types';
-import { useDragNode, UseDragNodeOptions } from './useDragNode';
-import { useDropNode, UseDropNodeOptions } from './useDropNode';
-
-export interface UseDndNodeOptions
-  extends Pick<UseDropNodeOptions, 'id' | 'blockRef'>,
-    Pick<UseDragNodeOptions, 'type'> {
-  drag?: UseDragNodeOptions;
-  drop?: UseDropNodeOptions;
-  preview?: {
-    /**
-     * Whether to disable the preview.
-     */
-    disable?: boolean;
-
-    /**
-     * The reference to the preview element.
-     */
-    ref?: any;
-  };
-}
+import { WithPartial } from '@udecode/plate-core';
+import { useDndNode, UseDndNodeOptions } from './useDndNode';
+import { DRAG_ITEM_BLOCK } from './useDragBlock';
 
 /**
- * {@link useDragNode} and {@link useDropNode} hooks to drag and drop a node from the editor.
- * A default preview is used to show the node being dragged, which can be customized or removed.
- * Returns the drag ref and drop line direction.
+ * {@link useDndNode}
  */
-export const useDndNode = ({
-  id,
-  type,
-  blockRef,
-  preview: previewOptions = {},
-  drag: dragOptions,
-  drop: dropOptions,
-}: UseDndNodeOptions) => {
-  const editor = useEditorRef();
-
-  const [dropLine, setDropLine] = useState<DropLineDirection>('');
-
-  const [{ isDragging }, dragRef, preview] = useDragNode(editor, {
-    id,
-    type,
-    ...dragOptions,
+export const useDndBlock = (options: WithPartial<UseDndNodeOptions, 'type'>) =>
+  useDndNode({
+    type: DRAG_ITEM_BLOCK,
+    ...options,
   });
-  const [{ isOver }, drop] = useDropNode(editor, {
-    accept: type,
-    id,
-    blockRef,
-    dropLine,
-    onChangeDropLine: setDropLine,
-    ...dropOptions,
-  });
-
-  if (previewOptions.disable) {
-    drop(blockRef);
-    preview(getEmptyImage(), { captureDraggingState: true });
-  } else if (previewOptions.ref) {
-    drop(blockRef);
-    preview(previewOptions.ref);
-  } else {
-    preview(drop(blockRef));
-  }
-
-  if (!isOver && dropLine) {
-    setDropLine('');
-  }
-
-  return {
-    isDragging,
-    isOver,
-    dropLine,
-    dragRef,
-  };
-};
-
-export const useDndBlock = (options: UseDndNodeOptions) => useDndNode(options);
