+++ packages/ui/dnd/src/hooks/useDropBlock.ts
@@ -3,9 +3,9 @@ import { DRAG_ITEM_BLOCK } from './useDragBlock';
 import { useDropNode, UseDropNodeOptions } from './useDropNode';
 
 /**
- * @deprecated use {@link useDropNode}
+ * {@link useDropNode}
  */
-export const useDropBlockOnEditor = <V extends Value>(
+export const useDropBlock = <V extends Value>(
   editor: TReactEditor<V>,
   options: Omit<UseDropNodeOptions, 'accept'>
 ) => useDropNode(editor, { accept: DRAG_ITEM_BLOCK, ...options });
