+++ packages/ui/dnd/src/hooks/useDropNode.ts
@@ -1,24 +1,16 @@
 import { DropTargetMonitor, useDrop } from 'react-dnd';
 import { DropTargetHookSpec } from 'react-dnd/src/hooks/types';
-import {
-  collapseSelection,
-  findNode,
-  focusEditor,
-  isExpanded,
-  moveNodes,
-  TReactEditor,
-  Value,
-} from '@udecode/plate-core';
-import { Path } from 'slate';
+import { TReactEditor, Value } from '@udecode/plate-core';
+import { onDropNode } from '../transforms/onDropNode';
+import { onHoverNode } from '../transforms/onHoverNode';
 import { DragItemNode, DropLineDirection } from '../types';
-import { getHoverDirection, getNewDirection } from '../utils/index';
 
 export interface UseDropNodeOptions
   extends DropTargetHookSpec<DragItemNode, unknown, { isOver: boolean }> {
   /**
-   * The reference to the block being dragged.
+   * The reference to the node being dragged.
    */
-  blockRef: any;
+  nodeRef: any;
 
   /**
    * Id of the node.
@@ -55,68 +47,24 @@ export interface UseDropNodeOptions
  */
 export const useDropNode = <V extends Value>(
   editor: TReactEditor<V>,
-  { blockRef, id, dropLine, onChangeDropLine, ...options }: UseDropNodeOptions
+  { nodeRef, id, dropLine, onChangeDropLine, ...options }: UseDropNodeOptions
 ) => {
   return useDrop<DragItemNode, unknown, { isOver: boolean }>({
     drop: (dragItem, monitor) => {
-      const direction = getHoverDirection(dragItem, monitor, blockRef, id);
-      if (!direction) return;
-
-      const dragEntry = findNode(editor, {
-        at: [],
-        match: { id: dragItem.id },
-      });
-      if (!dragEntry) return;
-      const [, dragPath] = dragEntry;
-
-      focusEditor(editor);
-
-      let dropPath: Path | undefined;
-      if (direction === 'bottom') {
-        dropPath = findNode(editor, { at: [], match: { id } })?.[1];
-        if (!dropPath) return;
-
-        if (Path.equals(dragPath, Path.next(dropPath))) return;
-      }
-
-      if (direction === 'top') {
-        const nodePath = findNode(editor, { at: [], match: { id } })?.[1];
-
-        if (!nodePath) return;
-        dropPath = [
-          ...nodePath.slice(0, -1),
-          nodePath[nodePath.length - 1] - 1,
-        ];
-
-        if (Path.equals(dragPath, dropPath)) return;
-      }
-
-      if (direction) {
-        const _dropPath = dropPath as Path;
-
-        const before =
-          Path.isBefore(dragPath, _dropPath) &&
-          Path.isSibling(dragPath, _dropPath);
-        const to = before ? _dropPath : Path.next(_dropPath);
-
-        moveNodes(editor, {
-          at: dragPath,
-          to,
-        });
-      }
+      onDropNode(editor, { nodeRef, id, dragItem, monitor });
     },
     collect: (monitor) => ({
       isOver: monitor.isOver(),
     }),
     hover(item: DragItemNode, monitor: DropTargetMonitor) {
-      const direction = getHoverDirection(item, monitor, blockRef, id);
-      const dropLineDir = getNewDirection(dropLine, direction);
-      if (dropLineDir) onChangeDropLine(dropLineDir);
-
-      if (direction && isExpanded(editor.selection)) {
-        focusEditor(editor);
-        collapseSelection(editor);
-      }
+      onHoverNode(editor, {
+        nodeRef,
+        id,
+        dropLine,
+        onChangeDropLine,
+        dragItem: item,
+        monitor,
+      });
     },
     ...options,
   });
