+++ packages/ui/dnd/src/index.ts
@@ -6,4 +6,6 @@ export * from './createDndPlugin';
 export * from './types';
 export * from './components/index';
 export * from './hooks/index';
+export * from './queries/index';
+export * from './transforms/index';
 export * from './utils/index';
