+++ packages/ui/dnd/src/utils/index.ts
@@ -2,11 +2,5 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './focusBlockStartById';
-export * from './getBlocksWithId';
 export * from './getHoverDirection';
 export * from './getNewDirection';
-export * from './getNodesRange';
-export * from './removeBlocksAndFocus';
-export * from './selectBlockById';
-export * from './selectBlocksBySelectionOrId';
